Referee: Add uptodate to the should_own
When files in A are synced with B and a new file is created in A,
the Referee was not adding A in the new owners. Now the files that
are currently up-to-date are added by default.
This could cause a problem if the rule just changed, we should do
some experiments.