Don't wait for Router and Dealer in test_changes. Fix #22.
We were waiting for rep1's router and dealer to start. This was
causing a timeout after 5s, so all the transfers had been emited
in less than 5s.
This is due to the fact that the router and the dealer only start
when all the changes have been emited.