Remove TryAgain exception in the Plug
The TryAgain exception was allowing the drivers to say to the
Plug to call the handler again later.
However, the implementation could block Onitu for a long time.

The current implementation wasn't very helpful. A better solution
can be found, with different behaviour wether it's during a
transfer, a read, a deletion, etc…

This does not break anything as no driver was using it yet.