Use ZMQ to notify new events. Fix #53
With Escalator, in order to detect new events, we were making
a request every 0.1s in the Referee and the Dealer.

Now, events are still stored in the DB, but when a new event is
storred, a notification is sent to the Dealer or the Referee, so
we don't need to do periodic requests anymore.