Detect the deletions on Dropbox (v1).
The metadata system on Dropbox allows us to detect deletions. When
we get deltas of Dropbox, files with None metadata have been deleted.
One could think that's a problem because the case-correct filename is
contained in this metadata. But Dropbox backups it. So if we request
this file metadata (with the case-incorrect filename), we can still
retrieve it. From it, we can recover the 'true' filename and notify
the plug thanks to it.

- Add some functions in the test Dropbox driver