Rewrite Onitu startup. Needs a patched Circus.
This commit should remove most zombies issues (cf #101).
We now use the loop provided by Circus instead of
creating a new one. This allows us to use Circus' default
exit system.
The motivation behind using our own loop was to be able
to call a coroutine after the starting of the watchers.
This has been made possible by a patch to Circus which
should hopefully be integrated to the main release.

We now don't use use add_callback to add the watchers
but yield, which should prevent any conflict error.

This commit also fixes some issues when Onitu is given
a wrong setup file.