Plug: Fix unicode error with Thread.name
The Router and the Dealer were inheriting from Thread
and overriding the Thread's name via self.name. But
a Thread can't have an unicode name, so it was causing
issues with the lasts changes.

Now the Router and the Dealer are objects and their run
method is started in a new Thread by the Plug.