Evernote: Change upload_file to manage Evernote markup enclosing.
When creating an Evernote note on Onitu side one may omit the
ENML boilerplate markup every note is enclosed into. In such cases
Evernote raises specific errors the upload_file now tries to handle
by enclosing the given content with ENML default markup before trying
to upload again.

- Removed the ".html" appending to filenames due to appearing of
filename inconsistencies, e.g. if we create a file without it on Onitu
it gets uploaded on Evernote with the name as is, but if we change it
on evernote, Onitu creates a new file `name'.html, and that's wrong.