Evernote: fix delete_file handler for resource files.
When deleting a file on Evernote it is whether a note or a resource.
We have to deal with the special case of resources because there's
no dedicated API call to delete them. So if the deleted file is a
resource, we have to get the note it's embedded in, and manually
remove it from this note's resources list (considered best practice).
It leaves the problem of a broken resource reference in the note, but
AFAIK there's no better way as of today.

- Clean useless imports