Fix the "handlers don't get called" issue. Bidirectional transfer OK
The problem was we didn't set metadata.size on new photo detection
on Flickr. The bigger problem behind that is that Flickr gives us no
way to know the real size in bytes of its files. Our somewhat dirty
solution to this is to set a TEMPORARY metadata.size of 1 to trigger
get_file if other drivers are interested, and then update the size
in get_file, with the length of the data chunk.

- get_file, upload_file, delete_file handlers are OK

- Add a variety of functions in FlickrHandler class to easily manage
handlers needs, particularly the upload_file which has a lot
of caveats to avoid. To list a few:
   o check if the photo already exist or not
   o create the photoset if it doesn't exist
   o (even trickier) Recreate the photoset if it doesn't exist ANYMORE
     (thus Onitu having an obsolete photoset ID)

- Modify the check changes thread to handle the obsolete photoset IDs