Evernote: upload_file works for updating notes.
upload_file now works, only for notes that have already been
downloaded from Evernote.

- Add some hash checking: A recurrent problem with Evernote is that
modifying a note's content also changes its resources USNs, even if
the resources' content hasn't changed. So now we store and compare
each resource body hash with the Evernote one during update handling.
If the hashes are equal, we don't call plug.update_file.

- Implement a similar protection for files in upload_file. Sometimes
we get a lot of calls to upload_file even when the file contents haven't
changed. So before we actually do the update, we compare the hash of the
content Onitu provided and the Evernote content hash: we skip the upload if
they're equal

- Modify update_note_resource to be able to specify if we want to update
the metadata size or not